Improve attachment loading blockymodel quad import fixes Some progress on blockymodel export

+++ src/blockymodel.ts
@@ -49,13 +49,62 @@ export function setupBlockymodelCodec(): Codec {
 				nodes: [],
 				lod: 'auto'
 			}
+			let id_count = 1;
+
+			let formatVector = (input: ArrayVector3) => ({
+				x: input[0],
+				y: input[1],
+				z: input[2],
+			} as IVector);
+
+			function compileNode(group: Group): BlockymodelNode | undefined {
+				let collection = Collection.all.find(c => c.contains(group));
+				if (collection && (!options.attachment || options.attachment == collection.uuid)) {
+					return;
+				}
+
+				let euler = Reusable.euler1.set(
+					Math.degToRad(group.rotation[0]),
+					Math.degToRad(group.rotation[1]),
+					Math.degToRad(group.rotation[2]),
+					group.scene_object.rotation.order
+				);
+				let quaternion = Reusable.quat1.setFromEuler(euler);
+				let orientation = {
+					x: quaternion.x,
+					y: quaternion.y,
+					z: quaternion.z,
+					w: quaternion.x,
+				}
+				let node: BlockymodelNode = {
+					id: id_count.toString(),
+					name: group.name,
+					position: formatVector(group.origin),
+					orientation,
+				}
+				id_count++;
+
+				let shape_count = 0;
+				for (let child of group.children){
+					if (child instanceof Cube && shape_count == 0) {
+
+					}
+				}
+
+				return node;
+			}
+			for (let group of Outliner.root) {
+				let compiled = group instanceof Group && compileNode(group);
+				if (compiled) model.nodes.push(compiled);
+			}
+
 			if (options.raw) {
 				return model;
 			} else {
 				return autoStringify(model);
 			}
 		},
-		parse(model: BlockymodelJSON, path: string, args?: {attachment?: boolean}) {
+		parse(model: BlockymodelJSON, path: string, args?: {attachment?: string}) {
 			function parseVector(vec: IVector, fallback: ArrayVector3 = [0, 0, 0]): ArrayVector3 | undefined {
 				if (!vec) return fallback;
 				return Object.values(vec).slice(0, 3) as ArrayVector3;
@@ -66,9 +115,9 @@ export function setupBlockymodelCodec(): Codec {
 				let quaternion = new THREE.Quaternion();
 				quaternion.set(node.orientation.x, node.orientation.y, node.orientation.z, node.orientation.w);
 				let rotation = new THREE.Euler().setFromQuaternion(quaternion.normalize(), 'ZYX');
-				let name = node.name.replace(/[-]/g, '_');
+				let name = node.name;
 				let offset = parseVector(node.shape.offset);
-				let position = parseVector(node.position);
+				let position = parseVector(node.position);				
 
 				let group = new Group({
 					name,
@@ -82,6 +131,12 @@ export function setupBlockymodelCodec(): Codec {
 				new_groups.push(group);
 				group.addTo(parent_group);
 
+				if (!parent_node && args.attachment) {
+					group.name = args.attachment + ':' + group.name,
+					group.origin = group.origin;
+					console.log(parent_group)
+
+				}
 				if (parent_group instanceof Group) {
 					let parent_geo_origin = parent_group.children.find(cube => cube instanceof Cube)?.origin ?? parent_group.origin;
 					if (parent_geo_origin) {
@@ -141,11 +196,11 @@ export function setupBlockymodelCodec(): Codec {
 							}
 							case '+X': {
 								cube.rotation[1] += 90;
-								switchIndices(cube.stretch, 2, 2);
+								switchIndices(cube.stretch, 0, 2);
 								break;
 							}
 							case '-X': {
-								cube.rotation[0] -= 90;
+								cube.rotation[1] -= 90;
 								switchIndices(cube.stretch, 0, 2);
 								break;
 							}
@@ -279,7 +334,8 @@ export function setupBlockymodelCodec(): Codec {
 				// Roots
 				let attachment_node: Group | undefined;
 				if (args.attachment && node.shape?.type == 'none' && Group.all.length) {
-					attachment_node = Group.all.find(g => g.name == node.name);
+					let node_name = node.name;
+					attachment_node = Group.all.find(g => g.name == node_name);
 				}
 				parseNode(node, null, attachment_node);
 			}
@@ -292,19 +348,24 @@ export function setupBlockymodelCodec(): Codec {
 			
 			const new_textures: Texture[] = [];
 			if (isApp && path) {
+				let project = Project;
 				let dirname = PathModule.dirname(path);
 				let fs = requireNativeModule('fs', {scope: PathModule.resolve(dirname, '..')});
 
 				let texture_files = fs.readdirSync(dirname);
 				for (let file_name of texture_files) {
 					if (file_name.match(/\.png$/i)) {
-						let texture = new Texture().fromPath(PathModule.join(dirname, file_name)).add(false, true);
+						let path = PathModule.join(dirname, file_name);
+						let texture = Texture.all.find(t => t.path == path);
+						if (!texture) {
+							texture = new Texture().fromPath(path).add(false, true);
+						}
 						new_textures.push(texture);
 					}
 				}
 				if (!args?.attachment) {
 					let listener = Blockbench.on('select_mode', ({mode}) => {
-						if (mode.id != 'animate') return;
+						if (mode.id != 'animate' || project != Project) return;
 						listener.delete();
 						try {
 							let anim_folders = fs.readdirSync(PathModule.resolve(dirname, '../Animations/'));
@@ -329,10 +390,10 @@ export function setupBlockymodelCodec(): Codec {
 			return {new_groups, new_textures};
 		}
 	})
-	let export_action = new Action('export_optifine_part', {
-		name: 'Export OptiFine Part',
-		description: 'Export a single part for an OptiFine model',
-		icon: 'icon-optifine_file',
+	let export_action = new Action('export_blockymodel', {
+		name: 'Export Hytale Blockymodel',
+		description: 'Export a blockymodel file',
+		icon: 'icon-hytale',
 		category: 'file',
 		condition: () => Format.id == 'hytale_model',
 		click: function () {
